+++ core/execution/spot/market.go
@@ -15,21 +15,6 @@
 
 package spot
 
-// Copyright (C) 2023 Gobalsky Labs Limited
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as
-// published by the Free Software Foundation, either version 3 of the
-// License, or (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
 import (
 	"context"
 	"errors"
@@ -96,15 +81,16 @@ type Market struct {
 	fee                           *fee.Engine
 	referralDiscountRewardService fee.ReferralDiscountRewardService
 	volumeDiscountService         fee.VolumeDiscountService
-	liquidity                     common.MarketLiquidityEngine
+	liquidity                     *common.MarketLiquidity
 	liquidityEngine               common.LiquidityEngine
 
 	// deps engines
 	collateral common.Collateral
 	banking    common.Banking
 
-	broker common.Broker
-	closed bool
+	broker               common.Broker
+	closed               bool
+	finalFeesDistributed bool
 
 	parties map[string]struct{}
 
@@ -150,6 +136,7 @@ type Market struct {
 	expiringStopOrders      *common.ExpiringOrders
 
 	minDuration time.Duration
+	epoch       types.Epoch
 }
 
 // NewMarket creates a new market using the market framework configuration and creates underlying engines.
@@ -306,7 +293,10 @@ func (m *Market) Update(ctx context.Context, config *types.Market) error {
 		for _, po := range peggedOrders {
 			order, err := m.matching.GetOrderByID(po)
 			if err != nil {
-				continue
+				order = m.peggedOrders.GetParkedByID(po)
+				if order == nil {
+					continue
+				}
 			}
 			if !num.UintZero().Mod(order.PeggedOrder.Offset, m.mkt.TickSize).IsZero() {
 				m.cancelOrder(ctx, order.Party, order.ID)
@@ -510,6 +500,7 @@ func (m *Market) UpdateMarketState(ctx context.Context, changes *types.MarketSta
 			m.as.EndGovernanceSuspensionAuction()
 			m.leaveAuction(ctx, m.timeService.GetTimeNow())
 		} else {
+			m.as.EndGovernanceSuspensionAuction()
 			if m.as.GetState().Trigger == types.AuctionTriggerOpening {
 				m.mkt.State = types.MarketStatePending
 				m.mkt.TradingMode = types.MarketTradingModeOpeningAuction
@@ -517,7 +508,6 @@ func (m *Market) UpdateMarketState(ctx context.Context, changes *types.MarketSta
 				m.mkt.State = types.MarketStateSuspended
 				m.mkt.TradingMode = types.MarketTradingModeMonitoringAuction
 			}
-			defer func() { m.idgen = nil }()
 			m.checkAuction(ctx, m.timeService.GetTimeNow(), m.idgen)
 			m.broker.Send(events.NewMarketUpdatedEvent(ctx, *m.mkt))
 		}
@@ -745,6 +735,15 @@ func (m *Market) closeCancelledMarket(ctx context.Context) error {
 func (m *Market) closeMarket(ctx context.Context) error {
 	// final distribution of liquidity fees
 	m.liquidity.OnMarketClosed(ctx, m.timeService.GetTimeNow())
+	// final distribution of liquidity fees
+	if !m.finalFeesDistributed {
+		if err := m.liquidity.AllocateFees(ctx); err != nil {
+			m.log.Panic("failed to allocate liquidity provision fees", logging.Error(err))
+		}
+
+		m.liquidity.OnEpochEnd(ctx, m.timeService.GetTimeNow(), m.epoch)
+		m.finalFeesDistributed = true
+	}
 	err := m.cleanMarketWithState(ctx, types.MarketStateClosed)
 	if err != nil {
 		return err
@@ -1581,6 +1580,14 @@ func (m *Market) handleConfirmation(ctx context.Context, conf *types.OrderConfir
 
 	if len(transfers) > 0 {
 		m.broker.Send(events.NewLedgerMovements(ctx, transfers))
+	}
+
+	if !m.as.InAuction() {
+		aggressor := conf.Order.Party
+		if quantum, err := m.collateral.GetAssetQuantum(m.quoteAsset); err == nil && !quantum.IsZero() {
+			n, _ := num.UintFromDecimal(tradedValue.ToDecimal().Div(quantum))
+			m.marketActivityTracker.RecordNotionalTakerVolume(m.mkt.ID, aggressor, n)
+		}
 	}
 
 	m.feeSplitter.AddTradeValue(tradedValue)
@@ -2581,7 +2588,8 @@ func (m *Market) getSuppliedStake() *num.Uint {
 func (m *Market) canTrade() bool {
 	return m.mkt.State == types.MarketStateActive ||
 		m.mkt.State == types.MarketStatePending ||
-		m.mkt.State == types.MarketStateSuspended
+		m.mkt.State == types.MarketStateSuspended ||
+		m.mkt.State == types.MarketStateSuspendedViaGovernance
 }
 
 // cleanupOnReject removes all resources created while the market was on PREPARED state.
@@ -3013,8 +3021,11 @@ func (m *Market) OnEpochEvent(ctx context.Context, epoch types.Epoch) {
 	if epoch.Action == vega.EpochAction_EPOCH_ACTION_START {
 		m.liquidity.UpdateSLAParameters(m.mkt.LiquiditySLAParams)
 		m.liquidity.OnEpochStart(ctx, m.timeService.GetTimeNow(), m.markPrice, m.midPrice(), m.getTargetStake(), m.positionFactor)
+		m.epoch = epoch
 	} else if epoch.Action == vega.EpochAction_EPOCH_ACTION_END {
-		m.liquidity.OnEpochEnd(ctx, m.timeService.GetTimeNow(), epoch)
+		if !m.finalFeesDistributed {
+			m.liquidity.OnEpochEnd(ctx, m.timeService.GetTimeNow(), epoch)
+		}
 		m.updateLiquidityFee(ctx)
 
 		m.banking.RegisterTradingFees(ctx, m.quoteAsset, m.fee.TotalTradingFeesPerParty())
@@ -3029,6 +3040,7 @@ func (m *Market) OnEpochEvent(ctx context.Context, epoch types.Epoch) {
 }
 
 func (m *Market) OnEpochRestore(ctx context.Context, epoch types.Epoch) {
+	m.epoch = epoch
 	m.liquidityEngine.OnEpochRestore(epoch)
 }